refactor(frontend): split AgentConversation into hooks and views
AgentConversation.tsx: 2692 -> 679 lines, across 22 modules.

The seam that made this possible: four mutable scroll refs (stick,
armBottom, animateBottom, programmaticScroll) were written from five call
sites and read by two sets of effects. useScrollIntent owns them behind
three verbs — armGlide() on submit, armJump() on restore/adopt, follow()
for a send that leaves now — so producers never touch raw refs and the two
scroll engines can be lifted out whole.

  assets/      conversationLayout, runError, messageParts, scrollGeometry
  hooks/       useScrollIntent, useTranscriptScroll, useVirtuosoTranscript,
               useAgentChatSession, useSessionHydration, useComposerDraft,
               useComposerAttachments, useVoiceComposer, useTurnInspector,
               useOnboardingChat, useFirstRunSeed
  components/  MessageRow, TurnActivity, AgentTurn, AgentTranscript,
               TranscriptPlaceholder, AgentComposerDock

AgentTurn is memoized and turn numbers are precomputed once per message
set — previously every settled turn re-rendered on each streamed token and
the turn lookup was an O(n^2) scan.

Split against current content so the voice-input and file-upload features
survive: useVoiceComposer is a new module for the recorder state (it has no
home in the pre-feature structure), and useComposerAttachments owns
uploadsEnabled plus bindDropTarget(isBlocked) — a predicate read at event
time, since the blocked-drop guard depends on the recorder which depends on
addFiles. Verified line-by-line against the upstream diff.
